(BSR)[API] chore: move then delete pcapi scheduled tasks
