+++ api/src/pcapi/core/providers/commands.py
@@ -1,3 +1,4 @@
+import datetime
 import logging
 from collections.abc import Generator
 from contextlib import contextmanager
@@ -7,11 +8,15 @@ import click
 
 import pcapi.core.providers.models as providers_models
 import pcapi.core.providers.repository as providers_repository
+import pcapi.utils.cron as cron_decorators
 from pcapi.core.providers import allocine
 from pcapi.local_providers import provider_manager
 from pcapi.models import db
+from pcapi.models.feature import FeatureToggle
 from pcapi.utils.blueprint import Blueprint
 
+from .titelive_book_search import TiteliveBookSearch
+from .titelive_music_search import TiteliveMusicSearch
 from .titelive_utils import generate_titelive_gtl_from_file
 
 
@@ -102,8 +107,114 @@ def update_providables_by_provider_id(provider_id: int, limit: int | None) -> No
     provider_manager.synchronize_venue_providers(venue_providers, limit)
 
 
+@blueprint.cli.command("synchronize_allocine_stocks")
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.SYNCHRONIZE_ALLOCINE)
+def synchronize_allocine_stocks() -> None:
+    """Launch AlloCine synchronization."""
+    allocine_stocks_provider_id = providers_repository.get_provider_by_local_class("AllocineStocks").id
+    venue_providers = providers_repository.get_active_venue_providers_by_provider(allocine_stocks_provider_id)
+    provider_manager.synchronize_venue_providers(venue_providers)
+
+
+@blueprint.cli.command("synchronize_cine_office_stocks")
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.ENABLE_CDS_IMPLEMENTATION)
+def synchronize_cine_office_stocks() -> None:
+    """Launch Cine Office synchronization."""
+    cine_office_stocks_provider_id = providers_repository.get_provider_by_local_class("CDSStocks").id
+    venue_providers = providers_repository.get_active_venue_providers_by_provider(cine_office_stocks_provider_id)
+    provider_manager.synchronize_venue_providers(venue_providers)
+
+
+@blueprint.cli.command("synchronize_boost_stocks")
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.ENABLE_BOOST_API_INTEGRATION)
+def synchronize_boost_stocks() -> None:
+    """Launch Boost synchronization."""
+    boost_stocks_provider_id = providers_repository.get_provider_by_local_class("BoostStocks").id
+    venue_providers = providers_repository.get_active_venue_providers_by_provider(boost_stocks_provider_id)
+    provider_manager.synchronize_venue_providers(venue_providers)
+
+
+@blueprint.cli.command("synchronize_cgr_stocks")
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.ENABLE_CGR_INTEGRATION)
+def synchronize_cgr_stocks() -> None:
+    """Launch CGR synchronization."""
+    cgr_stocks_provider_id = providers_repository.get_provider_by_local_class("CGRStocks").id
+    venue_providers = providers_repository.get_active_venue_providers_by_provider(cgr_stocks_provider_id)
+    provider_manager.synchronize_venue_providers(venue_providers)
+
+
+@blueprint.cli.command("synchronize_ems_stocks")
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.ENABLE_EMS_INTEGRATION)
+def synchronize_ems_stocks_on_schedule() -> None:
+    """Launch EMS synchronization"""
+    provider_manager.synchronize_ems_venue_providers(from_last_version=True)
+
+
 @blueprint.cli.command("update_gtl")
 @click.option("-f", "--file", required=True, help="CSV extract of GTL_2023.xlsx with tab as separator", type=str)
 def update_gtl(file: str) -> None:
     generate_titelive_gtl_from_file(file)
     # TODO we can later automatically reindex only the offers for which the gtl changed
+
+
+@blueprint.cli.command("synchronize_titelive_music_products")
+@click.option(
+    "--from-date",
+    help="sync music products that were modified after from_date (YYYY-MM-DD)",
+    type=click.DateTime(),
+    default=None,
+)
+@click.option(
+    "--to-date",
+    help="sync music products that were modified before to_date (YYYY-MM-DD)",
+    type=click.DateTime(),
+    default=None,
+)
+@click.option(
+    "--from-page",
+    help="page to sync from, defaults to 1",
+    type=int,
+    default=1,
+)
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.SYNCHRONIZE_TITELIVE_API_MUSIC_PRODUCTS)
+def synchronize_titelive_music_products(
+    from_date: datetime.datetime | None, to_date: datetime.datetime | None, from_page: int
+) -> None:
+    TiteliveMusicSearch().synchronize_products(
+        from_date.date() if from_date else None, to_date.date() if to_date else None, from_page
+    )
+
+
+@blueprint.cli.command("synchronize_titelive_book_products")
+@click.option(
+    "--from-date",
+    help="sync book products that were modified after from_date (YYYY-MM-DD)",
+    type=click.DateTime(),
+    default=None,
+)
+@click.option(
+    "--to-date",
+    help="sync music products that were modified before to_date (YYYY-MM-DD)",
+    type=click.DateTime(),
+    default=None,
+)
+@click.option(
+    "--from-page",
+    help="page to sync from, defaults to 1",
+    type=int,
+    default=1,
+)
+@cron_decorators.log_cron_with_transaction
+@cron_decorators.cron_require_feature(FeatureToggle.SYNCHRONIZE_TITELIVE_PRODUCTS)
+def synchronize_titelive_book_products(
+    from_date: datetime.datetime | None, to_date: datetime.datetime | None, from_page: int
+) -> None:
+    TiteliveBookSearch().synchronize_products(
+        from_date.date() if from_date else None, to_date.date() if to_date else None, from_page
+    )
