+++ api/src/pcapi/core/search/commands/indexation.py
@@ -7,12 +7,12 @@ import click
 import pcapi.core.educational.repository as collective_offers_repository
 import pcapi.core.offers.api as offers_api
 import pcapi.core.offers.repository as offers_repository
+import pcapi.utils.cron as cron_decorators
 from pcapi import settings
 from pcapi.core import search
 from pcapi.core.educational.api.offer import unindex_expired_or_archived_collective_offers_template
 from pcapi.core.offerers import api as offerers_api
 from pcapi.core.search import staging_indexation
-from pcapi.scheduled_tasks.decorators import log_cron_with_transaction
 from pcapi.utils.blueprint import Blueprint
 from pcapi.utils.chunks import get_chunks
 
@@ -22,28 +22,28 @@ logger = logging.getLogger(__name__)
 
 
 @blueprint.cli.command("index_offers_in_algolia_by_offer")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_offers_in_algolia_by_offer() -> None:
     """Pop offers from indexation queue and reindex them."""
     search.index_offers_in_queue()
 
 
 @blueprint.cli.command("index_offers_in_algolia_by_venue")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_offers_in_algolia_by_venue() -> None:
     """Pop venues from indexation queue and reindex their offers."""
     search.index_offers_of_venues_in_queue()
 
 
 @blueprint.cli.command("index_collective_offer_templates")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_collective_offer_templates() -> None:
     """Pop collective offers template from indexation queue and reindex them."""
     search.index_collective_offers_templates_in_queue()
 
 
 @blueprint.cli.command("delete_expired_offers_in_algolia")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def delete_expired_offers_in_algolia() -> None:
     """Unindex offers that have expired.
 
@@ -55,35 +55,35 @@ def delete_expired_offers_in_algolia() -> None:
 
 
 @blueprint.cli.command("delete_expired_collective_offers_template_in_algolia")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def delete_expired_collective_offers_template_in_algolia() -> None:
     """Unindex collective offers template that have expired or are archived."""
     unindex_expired_or_archived_collective_offers_template()
 
 
 @blueprint.cli.command("index_offers_in_error_in_algolia_by_offer")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_offers_in_error_in_algolia_by_offer() -> None:
     """Pop offers from the error queue and reindex them."""
     search.index_offers_in_queue(from_error_queue=True)
 
 
 @blueprint.cli.command("index_collective_offers_templates_in_error")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_collective_offers_templates_in_error() -> None:
     """Pop collective offers template from the error queue and reindex them."""
     search.index_collective_offers_templates_in_queue(from_error_queue=True)
 
 
 @blueprint.cli.command("index_venues")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_venues() -> None:
     """Pop venues from indexation queue and reindex them."""
     search.index_venues_in_queue()
 
 
 @blueprint.cli.command("index_venues_in_error")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def index_venues_in_error() -> None:
     """Pop venues from the error queue and reindex them."""
     search.index_venues_in_queue(from_error_queue=True)
@@ -184,7 +184,7 @@ def _reindex_all_venues(algolia_batch_size: int, max_venues: int) -> None:
 
 
 @blueprint.cli.command("update_products_booking_count_and_reindex_offers")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def update_products_booking_count_and_reindex_offers() -> None:
     """
     update last 30 days booking count for all products,
@@ -212,6 +212,6 @@ def index_offers_staging(clear: bool) -> None:
 
 
 @blueprint.cli.command("clean_indexation_processing_queues")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def clean_indexation_processing_queues() -> None:
     search.clean_processing_queues()
