+++ api/src/pcapi/core/subscription/commands.py
@@ -1,7 +1,7 @@
 import logging
 
+import pcapi.utils.cron as cron_decorators
 from pcapi.core.mails.transactional.users.ubble import reminder_emails
-from pcapi.scheduled_tasks.decorators import log_cron_with_transaction
 from pcapi.utils.blueprint import Blueprint
 
 
@@ -10,6 +10,6 @@ logger = logging.getLogger(__name__)
 
 
 @blueprint.cli.command("send_beneficiary_subscription_reminders")
-@log_cron_with_transaction
+@cron_decorators.log_cron_with_transaction
 def send_beneficiary_subscription_reminders() -> None:
     reminder_emails.send_reminders()
