--- api/src/pcapi/scheduled_tasks/CONTRIBUTING.md
@@ -1,3 +1,0 @@
-# Package `scheduled_tasks`
-
-Ce package est obsolète. Les commandes doivent être placées dans les fichiers `core/*/commands.py`.
