+++ api/src/pcapi/scripts/install.py
@@ -14,6 +14,7 @@ def install_commands(app: flask.Flask) -> None:
         "pcapi.core.external.commands",
         "pcapi.core.external_bookings.commands",
         "pcapi.core.finance.commands",
+        "pcapi.core.fraud.commands",
         "pcapi.core.geography.commands",
         "pcapi.core.offerers.commands",
         "pcapi.core.offers.commands",
@@ -23,9 +24,6 @@ def install_commands(app: flask.Flask) -> None:
         "pcapi.core.search.commands.settings",
         "pcapi.core.subscription.commands",
         "pcapi.core.users.commands",
-        "pcapi.scheduled_tasks.commands",
-        "pcapi.scheduled_tasks.offerer_stats_commands",
-        "pcapi.scheduled_tasks.titelive_commands",
         "pcapi.scripts.backoffice_users.add_permissions_to_staging_specific_roles",
         "pcapi.scripts.beneficiary.import_test_users",
         "pcapi.scripts.check_pre_migrations",
@@ -35,7 +33,6 @@ def install_commands(app: flask.Flask) -> None:
         "pcapi.scripts.generate_public_api_openapi_json",
         "pcapi.scripts.generate_native_api_openapi_json",
         "pcapi.scripts.sandbox",
-        "pcapi.scripts.ubble_archive_past_identifications",
         "pcapi.scripts.move_offer.move_batch_offer",
         "pcapi.utils.db",
         "pcapi.utils.human_ids",
